Add Mapping route fixes

+++ server/src/services/mapping/configCompanyServices.ts
@@ -1,5 +1,5 @@
-import { prisma } from "../prismaClient";
-import { NotFoundError } from "../utils/errors";
+import { prisma } from "../../prismaClient";
+import { NotFoundError } from "../../utils/errors";
 
 interface createConfigInput {
   companyId: string
@@ -9,37 +9,35 @@ interface createConfigInput {
   }[]
 }
 
+function normalizeAccountingAccount(value: string | number): string {
+
+  let str = String(value)
+
+  // Remove all points
+  str = str.replace(/\D/g, "")
+
+  if (str.length < 10) {
+    str = str.padEnd(10, "0")
+  }
+
+  return str
+}
+
 // Create Config with json
 export const createConfig = async (data: createConfigInput) => {
 
-  const company = await prisma.company.findUnique({where: { id: data.companyId }})
+  const company = await prisma.company.findUnique({ where: { id: data.companyId } })
   if (!company) throw new NotFoundError("Empresa com este ID nao existe no banco de dados!")
-
-  const configsOfCompanies = await prisma.configCompany.findMany({where: { companyId: data.companyId }})
-  if (configsOfCompanies) throw new NotFoundError("Já existe Configurações para essa empresa!")
 
   const config = await prisma.configCompany.createMany({
     data: data.configs.map((row) => ({
       companyId: data.companyId,
-      accountingAccount: row.accountingAccount,
+      accountingAccount: normalizeAccountingAccount(row.accountingAccount),
       accountName: row.accountName
     })),
   });
 
   return config
-}
-
-
-// List all configs
-export const listConfigCompanies = async () => {
-
-  const configs = await prisma.configCompany.findMany({
-    include: {
-      company: true
-    }
-  })
-
-  return configs
 }
 
 
@@ -52,9 +50,6 @@ export const listConfigCompany = async (companyId: string) => {
   const configs = await prisma.configCompany.findMany({
     where: {
       companyId
-    },
-    include: {
-      company: true
     }
   })
 
@@ -67,6 +62,13 @@ export const updateConfigCompany = async (data: createConfigInput) => {
   const company = await prisma.company.findUnique({ where: { id: data.companyId } })
   if (!company) throw new NotFoundError("Empresa com este ID não existe no banco de dados!")
 
+  const accounts = data.configs.map(c => c.accountingAccount)
+  const duplicates = accounts.filter((item, idx) => accounts.indexOf(item) !== idx)
+
+  if (duplicates.length > 0) {
+    throw new NotFoundError(`Existe contas com o mesmo codigo na lista, codigos: ${[...new Set(duplicates)].join(', ')}`)
+  }
+
   // delete all current configs of company
   await prisma.configCompany.deleteMany({
     where: { companyId: data.companyId }
@@ -76,7 +78,7 @@ export const updateConfigCompany = async (data: createConfigInput) => {
   const configs = await prisma.configCompany.createMany({
     data: data.configs.map((row) => ({
       companyId: data.companyId,
-      accountingAccount: row.accountingAccount,
+      accountingAccount: normalizeAccountingAccount(row.accountingAccount),
       accountName: row.accountName,
     }))
   })
